Adjusted solveModel output and calcLogLikImproved input
Adjusted so that the 'expected observation' is a field of the sol structure output by solvreModel, rather than a second output variable.
And calcLogLikImproved now that the initial condition and lower and upper bound for the parameter(s) to be optimised (e.g. pOObs) as inputs rather then being hard coded.
ThetaMLEImproved changed so that it inlcudes the value of the parameter(s) to be optimised (and so has consistent format with ThetaMLE)