Generalisable transform function
Code now set up to include model-specific transform functions encoding the effect of the parameter that is optimised within the loop. For the SEIR model this is just a multiplication of the parameter (pObs) with the expected observation under the model solution with pObs=1. For other models it could be something else, and this can be specified via a function of the form transformSolutionMODEL()