Dealing with CIs outside interval
CIs with boundary outside profiling interval are now recorded as being on the edge of the interval, to enable (conservative) calculation of coverage properties.
Other minor improvement to how GS options are specified.
Threshold value is now specified as negative (previously positive) and is calculated directly from the chi-squared distribution function