Update findCIs.m
Update to findCIs.m so that if all points to one side of the MLE are below threshold (i.e. relevant output of find command is empty), the CI boundary is set to the MLE (as opposed to previously where it was set to the edge of the profiling interval).
Minor updates to profiling widths